+++ scripts/constants.js
@@ -1,8 +1,0 @@
-/**
- * Constants shared across build scripts.
- */
-module.exports = {
-    DEMO_ENVIRONMENTS: [ 'development', 'production' ],
-    LIB_ENVIRONMENTS: [ 'development', 'production' ],
-    LIB_TARGETS: [ 'es', 'umd' ]
-};