+++ src/components/sets/SetsManager.js
@@ -1,5 +1,4 @@
 import React from 'react';
-import { Tabs } from 'antd';
 import SetsManagerTab from './SetsManagerTab';
 import SetsManagerActionBar from './SetsManagerActionBar';
 import 'antd/es/tabs/style/index.css';
@@ -17,15 +16,17 @@ export default function SetsManager(props) {
 
   return (
     <div className="sets-manager">
+      {/*
       <Tabs type="editable-card" onEdit={onEdit}>
         {setsTree.tabRoots.map(tabRoot => (
           <TabPane tab={tabRoot.name} key={tabRoot.setKey}>
             <SetsManagerTab setsTree={setsTree} tabRoot={tabRoot} />
           </TabPane>
         ))}
-      </Tabs>
+        </Tabs>
 
       <SetsManagerActionBar {...props} />
+      */}
     </div>
   );
 }