+++ src/components/sets/SetsManagerActionBar.js
@@ -1,10 +1,10 @@
 import React from 'react';
-import { Icon } from 'antd';
+import Icon from '@ant-design/icons';
 import 'antd/es/icon/style/index.css';
 import PopoverMenu from './PopoverMenu';
-import { ReactComponent as SetUnionSVG } from '../../assets/sets/union.svg';
-import { ReactComponent as SetIntersectionSVG } from '../../assets/sets/intersection.svg';
-import { ReactComponent as SetComplementSVG } from '../../assets/sets/complement.svg';
+import SetUnionSVG from '../../assets/sets/union.svg';
+import SetIntersectionSVG from '../../assets/sets/intersection.svg';
+import SetComplementSVG from '../../assets/sets/complement.svg';
 
 import {
   tabularFileType, tabularFileExtension,