+++ src/components/sets/TreeNode.js
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import { Icon } from 'antd';
+import Icon from '@ant-design/icons';
 import { TreeNode as RcTreeNode } from 'rc-tree';
 import { getDataAndAria } from 'rc-tree/es/util';
 import classNames from 'classnames';
@@ -7,9 +7,9 @@ import PopoverMenu from './PopoverMenu';
 import PopoverColor from './PopoverColor';
 import { callbackOnKeyPress, range, levelNameFromIndex } from './utils';
 
-import { ReactComponent as EyeSVG } from '../../assets/tools/eye.svg';
-import { ReactComponent as PenSVG } from '../../assets/tools/pen.svg';
-import { ReactComponent as TrashSVG } from '../../assets/tools/trash.svg';
+import EyeSVG from '../../assets/tools/eye.svg';
+import PenSVG from '../../assets/tools/pen.svg';
+import TrashSVG from '../../assets/tools/trash.svg';
 
 function makeNodeViewMenuConfig(props) {
   const {